feat: using fk's

+++ Security/Domain/Repositories/IUserRepository.cs
@@ -7,5 +7,5 @@ public interface IUserRepository
 {
     Task<IEnumerable<User>> ListUsersAsync();
     Task AddAsync(User user);
-    Task<User> FindByUserIdAsync(int userId);
+    Task<User> FindByUserIdAsync(long userId);
 }
